# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: N-Tech C³ Engineering Intelligence Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Not found
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Evidence Vault
  /evidence:
    get:
      operationId: listEvidence
      tags: [evidence]
      summary: List evidence items
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: storyId
          in: query
          schema:
            type: ["number", "null"]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of evidence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Evidence"
    post:
      operationId: createEvidence
      tags: [evidence]
      summary: Create an evidence item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
  /evidence/{id}:
    get:
      operationId: getEvidence
      tags: [evidence]
      summary: Get an evidence item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Evidence item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
        "404":
          description: Not found
    patch:
      operationId: updateEvidence
      tags: [evidence]
      summary: Update an evidence item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidencePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Evidence"
        "404":
          description: Not found
    delete:
      operationId: deleteEvidence
      tags: [evidence]
      summary: Delete an evidence item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Assets
  /assets:
    get:
      operationId: listAssets
      tags: [assets]
      summary: List assets
      parameters:
        - name: type
          in: query
          schema:
            type: string
        - name: storyId
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: List of assets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Asset"
    post:
      operationId: createAsset
      tags: [assets]
      summary: Create an asset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
  /assets/{id}:
    get:
      operationId: getAsset
      tags: [assets]
      summary: Get an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Asset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
    patch:
      operationId: updateAsset
      tags: [assets]
      summary: Update an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Asset"
        "404":
          description: Not found
    delete:
      operationId: deleteAsset
      tags: [assets]
      summary: Delete an asset
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Knowledge Base
  /knowledge:
    get:
      operationId: listKnowledge
      tags: [knowledge]
      summary: List knowledge base pages
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of knowledge pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KnowledgePage"
    post:
      operationId: createKnowledge
      tags: [knowledge]
      summary: Create a knowledge page
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgePage"
  /knowledge/{id}:
    get:
      operationId: getKnowledge
      tags: [knowledge]
      summary: Get a knowledge page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Knowledge page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgePage"
        "404":
          description: Not found
    patch:
      operationId: updateKnowledge
      tags: [knowledge]
      summary: Update a knowledge page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KnowledgePatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgePage"
        "404":
          description: Not found
    delete:
      operationId: deleteKnowledge
      tags: [knowledge]
      summary: Delete a knowledge page
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # Templates
...
components:
  schemas:
...
    Campaign:
      type: object
      required: [id, title, status, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        objective:
          type: ["string", "null"]
        status:
          type: string
          enum: [Planning, Active, Paused, Completed, Archived]
        platforms:
          type: array
          items:
            type: string
        durationWeeks:
          type: ["number", "null"]
        storyCount:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CampaignInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        objective:
          type: string
        status:
          type: string
          enum: [Planning, Active, Paused, Completed, Archived]
        platforms:
          type: array
          items:
            type: string
        durationWeeks:
          type: number
    CampaignPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        objective:
          type: ["string", "null"]
        status:
          type: string
          enum: [Planning, Active, Paused, Completed, Archived]
        platforms:
          type: array
          items:
            type: string
        durationWeeks:
          type: ["number", "null"]

    # Evidence
    Evidence:
      type: object
      required: [id, title, type, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        type:
          type: string
          enum: [Screenshot, TerminalOutput, GitLog, Benchmark, Diagram, MeetingNotes, ResearchPDF, Image, Video, VoiceRecording, CodeSnippet, IssueReport, BuildLog, RepositoryAudit, Other]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        storyId:
          type: ["number", "null"]
        projectId:
          type: ["number", "null"]
        repository:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EvidenceInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [Screenshot, TerminalOutput, GitLog, Benchmark, Diagram, MeetingNotes, ResearchPDF, Image, Video, VoiceRecording, CodeSnippet, IssueReport, BuildLog, RepositoryAudit, Other]
        source:
          type: string
        notes:
          type: string
        content:
          type: string
        tags:
          type: array
          items:
            type: string
        storyId:
          type: ["number", "null"]
        projectId:
          type: ["number", "null"]
        repository:
          type: string
    EvidencePatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [Screenshot, TerminalOutput, GitLog, Benchmark, Diagram, MeetingNotes, ResearchPDF, Image, Video, VoiceRecording, CodeSnippet, IssueReport, BuildLog, RepositoryAudit, Other]
        source:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        storyId:
          type: ["number", "null"]
        projectId:
          type: ["number", "null"]
        repository:
          type: ["string", "null"]

    # Assets
    Asset:
      type: object
      required: [id, title, type, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        type:
          type: string
          enum: [Image, Icon, Illustration, Logo, Video, SVG, PDF, BrandAsset, ScreenRecording, Other]
        url:
          type: ["string", "null"]
        filePath:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        storyId:
          type: ["number", "null"]
        campaignId:
          type: ["number", "null"]
        projectId:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AssetInput:
      type: object
      required: [title, type]
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [Image, Icon, Illustration, Logo, Video, SVG, PDF, BrandAsset, ScreenRecording, Other]
        url:
          type: string
        filePath:
          type: string
        tags:
          type: array
          items:
            type: string
        storyId:
          type: ["number", "null"]
        campaignId:
          type: ["number", "null"]
        projectId:
          type: ["number", "null"]
    AssetPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        type:
          type: string
          enum: [Image, Icon, Illustration, Logo, Video, SVG, PDF, BrandAsset, ScreenRecording, Other]
        url:
          type: ["string", "null"]
        filePath:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        storyId:
          type: ["number", "null"]
        campaignId:
          type: ["number", "null"]
        projectId:
          type: ["number", "null"]

    # Knowledge Base
    KnowledgePage:
      type: object
      required: [id, title, createdAt, updatedAt]
      properties:
        id:
          type: number
        title:
          type: string
        content:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        linkedPageIds:
          type: array
          items:
            type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    KnowledgeInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: string
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        linkedPageIds:
          type: array
          items:
            type: number
    KnowledgePatch:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        content:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        linkedPageIds:
          type: array
          items:
            type: number

    # Templates
...