# lib/api-zod/src/generated/types/asset.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * N-Tech C³ Engineering Intelligence Operating System API
 * OpenAPI spec version: 0.1.0
 */
import type { AssetType } from './assetType';

export interface Asset {
  id: number;
  title: string;
  type: AssetType;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  filePath?: string | null;
  tags?: string[];
  /** @nullable */
  storyId?: number | null;
  /** @nullable */
  campaignId?: number | null;
  /** @nullable */
  projectId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}