# lib/api-zod/src/generated/types/knowledgePatch.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * N-Tech C³ Engineering Intelligence Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface KnowledgePatch {
  /** @minLength 1 */
  title?: string;
  /** @nullable */
  content?: string | null;
  /** @nullable */
  category?: string | null;
  tags?: string[];
  linkedPageIds?: number[];
}